pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentralert"
version = "0.1.0"
description = "Intelligent alert generation system for monitoring platforms using Claude AI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.39.0",
    "requests>=2.32.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
]

[project.scripts]
sentralert = "sentralert.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
